refactoring work: project-info --> sync-server/monitor/activity
- needed to support this functionality outside of the home base project
