+++ src/packages/project/package.json
@@ -27,8 +27,8 @@
     "@cocalc/project": "workspace:*",
     "@cocalc/sync": "workspace:*",
     "@cocalc/sync-client": "workspace:*",
-    "@cocalc/sync-server": "workspace:*",
     "@cocalc/sync-fs": "workspace:*",
+    "@cocalc/sync-server": "workspace:*",
     "@cocalc/terminal": "workspace:*",
     "@cocalc/util": "workspace:*",
     "@nteract/messaging": "^7.0.20",
@@ -41,7 +41,6 @@
     "compression": "^1.7.4",
     "daemonize-process": "^3.0.0",
     "debug": "^4.3.2",
-    "diskusage": "^1.1.3",
     "expect": "^26.6.2",
     "express": "^4.21.1",
     "express-rate-limit": "^7.4.0",
@@ -83,7 +82,9 @@
     "clean": "rm -rf dist"
   },
   "author": "SageMath, Inc.",
-  "contributors": ["William Stein <wstein@sagemath.com>"],
+  "contributors": [
+    "William Stein <wstein@sagemath.com>"
+  ],
   "license": "SEE LICENSE.md",
   "bugs": {
     "url": "https://github.com/sagemathinc/cocalc/issues"
