+++ src/packages/sync-server/monitor/README.md
@@ -1,0 +1,9 @@
+The modules here all collect and update tables that provide various
+info about the sync-server, which is relevant to the sync client.
+
+- process
+- disk usage
+- cpu usage
+- X11 ports
+
+etc
