+++ src/packages/sync-server/monitor/activity/project-info.ts
