+++ src/packages/sync-server/package.json
@@ -3,7 +3,8 @@
   "version": "0.1.0",
   "description": "CoCalc realtime synchronization framework -- server component",
   "exports": {
-    "./*": "./dist/*.js"
+    "./*": "./dist/*.js",
+    "./monitor/activity": "./dist/monitor/activity/index.js"
   },
   "scripts": {
     "clean": "rm -rf dist",
@@ -19,12 +20,14 @@
   "dependencies": {
     "@cocalc/backend": "workspace:*",
     "@cocalc/jupyter": "workspace:*",
-    "@cocalc/sync-server": "workspace:*",
     "@cocalc/sync": "workspace:*",
+    "@cocalc/sync-server": "workspace:*",
     "@cocalc/terminal": "workspace:*",
     "@cocalc/util": "workspace:*",
+    "awaiting": "^3.0.0",
+    "diskusage": "^1.1.3",
     "events": "3.3.0",
-    "awaiting": "^3.0.0"
+    "lodash": "^4.17.21"
   },
   "homepage": "https://github.com/sagemathinc/cocalc/tree/master/src/packages/sync-server",
   "repository": {
@@ -32,6 +35,7 @@
     "url": "https://github.com/sagemathinc/cocalc"
   },
   "devDependencies": {
+    "@types/lodash": "^4.14.202",
     "@types/node": "^18.16.14"
   }
 }
